React to review.
