+++ src/OrchardCore/OrchardCore/Modules/Extensions/ServiceCollectionExtensions.cs
@@ -1,6 +1,4 @@
-using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using Microsoft.AspNetCore.Authentication;
@@ -20,6 +18,7 @@ using OrchardCore.Environment.Extensions;
 using OrchardCore.Environment.Shell;
 using OrchardCore.Environment.Shell.Configuration;
 using OrchardCore.Environment.Shell.Descriptor.Models;
+using OrchardCore.Localization;
 using OrchardCore.Modules;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -252,27 +251,6 @@ namespace Microsoft.Extensions.DependencyInjection
 
                 services.Add(collection);
             });
-        }
-
-        private class NullStringLocalizerFactory : IStringLocalizerFactory
-        {
-            public IStringLocalizer Create(Type resourceSource) => NullStringLocalizer.Instance;
-            public IStringLocalizer Create(string baseName, string location) => NullStringLocalizer.Instance;
-        }
-
-        private class NullStringLocalizer : IStringLocalizer
-        {
-            public static NullStringLocalizer Instance { get; } = new NullStringLocalizer();
-
-            public LocalizedString this[string name] => new LocalizedString(name, name);
-
-            public LocalizedString this[string name, params object[] arguments]
-                => new LocalizedString(name, string.Format(name, arguments));
-
-            public IEnumerable<LocalizedString> GetAllStrings(bool includeParentCultures)
-                => Enumerable.Empty<LocalizedString>();
-
-            public IStringLocalizer WithCulture(CultureInfo culture) => Instance;
         }
     }
 }
